+++ src/types/index.tsx
@@ -1,10 +1,12 @@
+export type CreditCardPasswordType = {
+  first: string,
+  second: string
+};
+
 export type CreditCard = {
   number: string;
   expiry: string;
   owner?: string;
   cvc: string;
-  password: {
-    first: string;
-    second: string;
-  };
+  password: CreditCardPasswordType;
 };
